Update VSCode folder structure
