+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/CSharp/CSharpProjectedDocument.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { IProjectedDocument } from '../IProjectedDocument';
+import { IProjectedDocument } from '../Projection/IProjectedDocument';
 import { ServerTextChange } from '../RPC/ServerTextChange';
 import { getUriPath } from '../UriPaths';
 import * as vscode from '../vscodeAdapter';
