+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/CodeActions/RazorCodeAction.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
+import { SerializableWorkspaceEdit } from '../RPC/SerializableWorkspaceEdit';
 import { RazorCodeActionDataParams } from './RazorCodeActionDataParams';
 import { RazorCodeActionResolutionParams } from './RazorCodeActionResolutionParams';
-import { SerializableWorkspaceEdit } from './SerializableWorkspaceEdit';
 
 export interface RazorCodeAction {
     title: string;
