+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/CodeActions/RazorCodeActionRunner.ts
@@ -6,9 +6,9 @@
 import * as vscode from 'vscode';
 import { RazorLanguageServerClient } from '../RazorLanguageServerClient';
 import { RazorLogger } from '../RazorLogger';
-import { RazorCodeAction } from '../RPC/RazorCodeAction';
-import { RazorCodeActionResolutionParams } from '../RPC/RazorCodeActionResolutionParams';
 import { convertWorkspaceEditFromSerializable } from '../RPC/SerializableWorkspaceEdit';
+import { RazorCodeAction } from './RazorCodeAction';
+import { RazorCodeActionResolutionParams } from './RazorCodeActionResolutionParams';
 
 export class RazorCodeActionRunner {
     private static readonly codeActionResolveEndpoint = 'textDocument/codeActionResolve';
