+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/ColorPresentation/ColorPresentationHandler.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from 'vscode';
 import { RequestType } from 'vscode-languageclient';
-import { RazorDocumentManager } from '../RazorDocumentManager';
+import { RazorDocumentManager } from '../Document/RazorDocumentManager';
 import { RazorLanguageServerClient } from '../RazorLanguageServerClient';
 import { RazorLogger } from '../RazorLogger';
 import { convertTextEditToSerializable, SerializableTextEdit } from '../RPC/SerializableTextEdit';
