+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Document/IRazorDocumentChangeEvent.ts
