+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/DocumentColor/DocumentColorHandler.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from 'vscode';
 import { RequestType } from 'vscode-languageclient';
-import { RazorDocumentManager } from '../RazorDocumentManager';
+import { RazorDocumentManager } from '../Document/RazorDocumentManager';
 import { RazorLanguageServerClient } from '../RazorLanguageServerClient';
 import { RazorLogger } from '../RazorLogger';
 import { convertRangeToSerializable } from '../RPC/SerializableRange';
