+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/DocumentTelemetryListener.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { RazorDocumentChangeKind } from './RazorDocumentChangeKind';
-import { RazorDocumentManager } from './RazorDocumentManager';
+import { RazorDocumentChangeKind } from './Document/RazorDocumentChangeKind';
+import { RazorDocumentManager } from './Document/RazorDocumentManager';
 import { TelemetryReporter } from './TelemetryReporter';
 
 export function reportTelemetryForDocuments(
