+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Folding/SerializableFoldingRangeResponse.ts
