+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Html/HtmlProjectedDocument.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { IProjectedDocument } from '../IProjectedDocument';
+import * as vscode from 'vscode';
+import { IProjectedDocument } from '../Projection/IProjectedDocument';
 import { ServerTextChange } from '../RPC/ServerTextChange';
 import { getUriPath } from '../UriPaths';
-import * as vscode from '../vscodeAdapter';
 
 export class HtmlProjectedDocument implements IProjectedDocument {
     public readonly path: string;
