+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Html/HtmlTagCompletionProvider.ts
@@ -9,7 +9,7 @@ import {
     LanguageService as HtmlLanguageService,
 } from 'vscode-html-languageservice';
 import { TextDocument as ServiceTextDocument } from 'vscode-languageserver-textdocument';
-import { IRazorDocumentManager } from '../IRazorDocumentManager';
+import { IRazorDocumentManager } from '../Document/IRazorDocumentManager';
 import { RazorLanguage } from '../RazorLanguage';
 import { RazorLanguageServiceClient } from '../RazorLanguageServiceClient';
 import { LanguageKind } from '../RPC/LanguageKind';
