+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Html/RazorHtmlFeature.ts
@@ -4,8 +4,8 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as vscode from 'vscode';
+import { RazorDocumentManager } from '../Document/RazorDocumentManager';
 import { IEventEmitterFactory } from '../IEventEmitterFactory';
-import { RazorDocumentManager } from '../RazorDocumentManager';
 import { RazorLanguageServiceClient } from '../RazorLanguageServiceClient';
 import { RazorLogger } from '../RazorLogger';
 import { HtmlPreviewPanel } from './HtmlPreviewPanel';
