+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Projection/IProjectedDocument.ts
