+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Projection/ProjectionResult.ts
@@ -3,9 +3,11 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { SerializableTextEdit } from './SerializableTextEdit';
+import * as vscode from 'vscode';
+import { LanguageKind } from '../RPC/LanguageKind';
 
-export class RazorDocumentRangeFormattingResponse {
-    constructor(public readonly edits: SerializableTextEdit[]) {
-    }
+export interface ProjectionResult {
+    uri: vscode.Uri;
+    position: vscode.Position;
+    languageKind: LanguageKind;
 }
