+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/RazorLanguageFeatureBase.ts
@@ -4,9 +4,9 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as vscode from 'vscode';
-import { ProjectionResult } from './ProjectionResult';
-import { RazorDocumentManager } from './RazorDocumentManager';
-import { RazorDocumentSynchronizer } from './RazorDocumentSynchronizer';
+import { RazorDocumentManager } from './Document/RazorDocumentManager';
+import { RazorDocumentSynchronizer } from './Document/RazorDocumentSynchronizer';
+import { ProjectionResult } from './Projection/ProjectionResult';
 import { RazorLanguageServiceClient } from './RazorLanguageServiceClient';
 import { RazorLogger } from './RazorLogger';
 import { LanguageKind } from './RPC/LanguageKind';
