+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Rename/SerializableRenameDocument.ts
