+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/extension.ts
@@ -9,41 +9,41 @@ import { ExtensionContext } from 'vscode';
 import { BlazorDebugConfigurationProvider } from './BlazorDebug/BlazorDebugConfigurationProvider';
 import { CodeActionsHandler } from './CodeActions/CodeActionsHandler';
 import { RazorCodeActionRunner } from './CodeActions/RazorCodeActionRunner';
+import { RazorCodeLensProvider } from './CodeLens/RazorCodeLensProvider';
 import { ColorPresentationHandler } from './ColorPresentation/ColorPresentationHandler';
+import { ProvisionalCompletionOrchestrator } from './Completion/ProvisionalCompletionOrchestrator';
+import { RazorCompletionItemProvider } from './Completion/RazorCompletionItemProvider';
 import { listenToConfigurationChanges } from './ConfigurationChangeListener';
 import { RazorCSharpFeature } from './CSharp/RazorCSharpFeature';
+import { RazorDefinitionProvider } from './Definition/RazorDefinitionProvider';
 import { ReportIssueCommand } from './Diagnostics/ReportIssueCommand';
+import { RazorDocumentManager } from './Document/RazorDocumentManager';
+import { RazorDocumentSynchronizer } from './Document/RazorDocumentSynchronizer';
 import { DocumentColorHandler } from './DocumentColor/DocumentColorHandler';
 import { RazorDocumentHighlightProvider } from './DocumentHighlight/RazorDocumentHighlightProvider';
 import { reportTelemetryForDocuments } from './DocumentTelemetryListener';
 import { FoldingRangeHandler } from './Folding/FoldingRangeHandler';
 import { FormattingHandler } from './Formatting/FormattingHandler';
+import { RazorFormattingFeature } from './Formatting/RazorFormattingFeature';
 import { HostEventStream } from './HostEventStream';
+import { RazorHoverProvider } from './Hover/RazorHoverProvider';
 import { RazorHtmlFeature } from './Html/RazorHtmlFeature';
 import { IEventEmitterFactory } from './IEventEmitterFactory';
+import { RazorImplementationProvider } from './Implementation/RazorImplementationProvider';
 import { ProposedApisFeature } from './ProposedApisFeature';
-import { ProvisionalCompletionOrchestrator } from './ProvisionalCompletionOrchestrator';
-import { RazorCodeLensProvider } from './RazorCodeLensProvider';
-import { RazorCompletionItemProvider } from './RazorCompletionItemProvider';
 import { RazorCSharpLanguageMiddleware } from './RazorCSharpLanguageMiddleware';
-import { RazorDefinitionProvider } from './RazorDefinitionProvider';
-import { RazorDocumentManager } from './RazorDocumentManager';
-import { RazorDocumentSynchronizer } from './RazorDocumentSynchronizer';
-import { RazorFormattingFeature } from './RazorFormattingFeature';
-import { RazorHoverProvider } from './RazorHoverProvider';
-import { RazorImplementationProvider } from './RazorImplementationProvider';
 import { RazorLanguage } from './RazorLanguage';
 import { RazorLanguageConfiguration } from './RazorLanguageConfiguration';
 import { RazorLanguageServerClient } from './RazorLanguageServerClient';
 import { resolveRazorLanguageServerTrace } from './RazorLanguageServerTraceResolver';
 import { RazorLanguageServiceClient } from './RazorLanguageServiceClient';
 import { RazorLogger } from './RazorLogger';
-import { RazorReferenceProvider } from './RazorReferenceProvider';
-import { RazorRenameProvider } from './RazorRenameProvider';
 import { RazorServerReadyHandler } from './RazorServerReadyHandler';
-import { RazorSignatureHelpProvider } from './RazorSignatureHelpProvider';
+import { RazorReferenceProvider } from './Reference/RazorReferenceProvider';
+import { RazorRenameProvider } from './Rename/RazorRenameProvider';
 import { RazorDocumentSemanticTokensProvider } from './Semantic/RazorDocumentSemanticTokensProvider';
 import { SemanticTokensRangeHandler } from './Semantic/SemanticTokensRangeHandler';
+import { RazorSignatureHelpProvider } from './SignatureHelp/RazorSignatureHelpProvider';
 import { TelemetryReporter } from './TelemetryReporter';
 
 // We specifically need to take a reference to a particular instance of the vscode namespace,
